Fixing a lot

+++ src/PoViEmu.Expert/XIntel16x0B.cs
@@ -313,7 +313,7 @@ namespace PoViEmu.Expert
                 case 0xA2:
                     return new(pos, first, 4, O.or, [R.SP, R.BP.Plus(R.SI).Signed(s.NextShort(buff))],   [ second ]  );
                 case 0xA3:
-                    return new(pos, first, 4, O.or, [R.SP, R.BP.Plus(R.DI).Signed(s.NextByte(buff))],   [ second ]  );
+                    return new(pos, first, 4, O.or, [R.SP, R.BP.Plus(R.DI).Signed(s.NextShort(buff))],   [ second ]  );
                 case 0xA4:
                     return new(pos, first, 4, O.or, [R.SP, R.SI.Signed(s.NextShort(buff))],   [ second ]  );
                 case 0xA5:
